// src/app/enums/interests.ts
export enum Interest {
    Nature = "Nature et paysages (montagnes, plages, parcs naturels, lacs, etc.)",
    Culture = "Culture (musées, monuments, histoire, art, etc.)",
    Gastronomy = "Gastronomie (spécialités locales, vins, restaurants, etc.)",
    NightLife = "Vie nocturne (bars, clubs, concerts, festivals, etc.)",
    OutdoorActivities = "Sports et activités de plein air (randonnée, vélo, surf, escalade, ski, etc.)",
    WellnessRelaxation = "Bien-être et détente (spa, thermes, yoga, massages, etc.)",
    Shopping = "Shopping (marchés, boutiques, centres commerciaux, etc.)",
    SpecialEvents = "Événements spéciaux (carnavals, foires, festivals, etc.)",
    ChildrenFamily = "Famille et enfants (parcs d'attractions, zoos, aquariums, etc.)",
    ArchitectureDesign = "Architecture et design (bâtiments historiques, modernes, art urbain, etc.)",
}